hos-backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.16,<4.0
requests>=2.32,<3.0
django-cors-headers>=4.9,<5.0
